Implement piggy-backed credit returns, knocking 1/3 off the mailbox memory. Realtive to 48 hours ago, we now use 1/2 the memory for AM mailboxes.
The main "trick" (used in most other conduits already) is to use a struct
for the token, and to record in that struct whether or not a reply is
needed.  The 'need_reply' field is initially True for the token of a
Request, and is False for a Reply.  The AMReply functions will assert that
the field is True (catching multiple Replies and Reply-to-Reply behaviors)
and then set it to False.
The AM dispatcher code knows that if the 'need_reply' field is still True
after the Request handler is run, then a "hidden" credit return message is
needed for flow control.
The other "side" of this logic is that every Reply now returns the credit
which was obtained earlier by the corresponing Request.