Merge the 2 scratch buffers I had previously split (and enlarge).
Because multi-image all-to-all is the only case neededing 2 scratch buffers
I am reverting to keeping just 1 buffer and the all-to-all code partitions
it into 2 pieces when needed.  This allows the other multi-image collectives
to operate to twice the max size within the same total memory allocation.
The amount of space allocated (only for TEAM_ALL so far), has been enlarged
from 4K to 1M with this commit.