Fix a counting error reported by Valentin Petrov
The ssh-based spawner was incorrectly counting the children it spawned
during abnormal termination.  The result is that the counter could become
negative and the master process would never terminate (since it is waiting
for the count to become zero).  So, the job would hang on some abnormal
exit cases that the code WAS designed to handle properly.