Eliminate mutex lock/unlock from notify step in RDMADISSEM barrier.
By merging "phase" and "step" into a single "slot" variable we can do away
with the need to lock durring the notify step, and still read only valid
values in the kick function (which still needs the lock to prevent any
concurrent kickers).  The main role of the "step" and "phase" used them in
exactly the merge role (step << 1)|phase already.  So, this was a minor
change in bookkeeping only.

This makes a small (but measurable) difference in timings, but there is
still a BIG penalty for the presence of pollers threads in testbarrier-par.
I tried the few obvious ways to have a WAIT exclude the kick function, but
didn't see improvements (and sometimes things got worse).  This was only
tested with gemini-conduit, and perhaps excluding the pollers will be a
plus when measured on another conduit.