Commit (disabled) implementation of single-sender variant of the RDMADISSEM barrier.
When using the PSHM-heirarchical barrier there is a single designated representative
proc per each supernode which performs MOST of the communication.  However, by default
the last process per supernode to reach the Notify performs the first Put to get it
sent without delay.  When the single-sender logic is enbaled, the designated proc
will do ALL of the Puts even though this will often delay the first one.  Of course
this hasn't yielded any improvements worth noting in any testing so far.  However,
it is worth cosidering that this may be useful for a conduit with dynamic connections
which could possibly reduce the number of communication peers by enabling this code.

It is worth noting that I've also tried the variation in which the last process per
supernode to arrive at the Notify is "elected" to be the communication representative
for the current barrier.  However, for our benchmarks (which are timing numerous back-
to-back barriers) this is never a win (possibly because of the ping-ponging of one or
mode cachelines among the ever-changing representatives) and doesn't have a good
motivation behind it like reducing the number of comm pairs as single-sender can.
I *might* commit that logic too, but haven't yet merged it with this commit.