Commit the GASNet half of the attach hook logic.
This is a portion of the work Filip Blagojovic was engaged in before he left.
This commit is only the GASNet side of things and allows the client to get
a callback between creation/allocation of the GASNet segment and any
conduit-specific registration/pinning.  The purpose is to allow, for instance,
for a multi-threaded client on a NUMA platform to perform first-touch on
the appropriate portions of the segmement before memory registration.
That is exactly what the (not yet ready to commit) UPCR half of this work does.