Commit "merged" lin/log intra-node barrier code.
This allows run time selection (via env var) of the Linear vs Logarithmic
versions of the shared-memory barrier.
The default is currently Linear on all platforms.

Work is planned to hoepfully improve the Linear algorithm's performance based
on past work Dan did for Titanium.  So, the final "bake off" is not complete.