Commit initially disabled variant of Logarithmic intra-node barrier based on atomic SWAP.
Since we need to detect the 2nd arrival, the use of a SWAP (or any test-and-set or
test-and-clear for that matter) is a viable alternative to the dec-and-test being
used currently.  The swap-based variant has the (very small) advantage of not requiring
any reset because we use the PHASE in the value we ae swapping in.
Test data collected with the swap variant shows only a VERY small advantage, if any,
Additionally, when the data gets noisy in the high core count tests, it looks
(visually, not necessarily statistically) as if the SWAP variant is more sensitve
(higher variance) than the dec-and-test code.
So, this is disabled for now.