Use a union to organize the per-node shared-memory barrier data.
As indicated in a comment, this is in preparation for a possible/probable
variant of the linear barrier code in which we pack the entire struct into
a single 64-bit word.  On platforms where 64-bit read and write are free
of word-tearing, we then get to eliminate the rmb/wmb pairing currently
used to ensure that the value/flags we read are up-to-date with respect
to the phase value we've read.