Replace the Logarithmic shared-memory barrier code.
Following up on another of Dan's ideas in the intra-node barrier email thread,
this commit generalizes the NON-atomic barrier code (which already replaced
the Linear algorithm) to be tree-structured.
This REPLACES the atomic-based Logarithmic barrier, since the performance is
seen to be uniformly better.
For the moment, the default remains RADIX=0 which makes a "flat tree"
in which the rank==0 node is the parent of all other nodes (and this
is EXACTLY the prior Linear algorithm)