Revise my solution for conduit-threads which require threaddata.
A conduit can now indicate if (and how many) conduit-threads may potentially
call the extended-api (incl. via progress fns) and therefore need slots in
the threaddata table.  The value of GASNETI_MAX_THREADS for a SEQ build is
adjusted accordingly, and all logic that was recenty changed to use
"#if GASNETI_THREADS" to enable multi-threaded handling of threaddata has
been changed to now use "#if GASNETI_MAX_THREADS > 1".

This means that one can have conduit threads (such as the connection thread
in vapi/ibv-conduit) which require true mutexes and thread-safty for things
like the gasnet malloc wrappers, but which don't require a threaddata for
themselves.  Therefore, a SEQ build w/ ONLY such a thread will use the ONE
global threaddata without the table and without (possibly expensive)
dynamic lookups via pthread_getspecific().  In this respect, this commit
returns us to the state of things a few days ago, while preserving the
required change to introduce a threaddata for threads with run any progress
functions (as the vapi/ibv-conduit receive thread does).