Recode PAMIDISSEM barrier and make it the default on BG/Q.
By switching from the reduction-style expression to the more standard way
of expressing the matching, we reduce the network message size from 20
bytes to just 10 (though in theory the 20 could have been packed down to
fit in just 14).  This results in measurable performance improvement for
all but the SMT4 cases; when using all 4 threads per core the additional
compute work done in the "standard" expression of the name matching shows
its cost relative to the very simple MAX().  The difference is around
0.5% and would be unmeasurable if not for the amazing reproducability
on BG/Q.  For 1 or 2 threads per core this commit produces an improvement.

Additonally, this is now the default on BG/Q where is is as much as 50%
faster than the PAMIALLREDUCE barrier.  The previous measurements that
found PAMIDISSEM to be comparable to AMDISSEM were, in fact, flawed.
The results were "comparable" because I was actually measuring AMDISSEM!
Not until some fixes committed overnight did the PAMIDISSEM barrier work
at all, but due to a flawed _init routine, it had never actually enabled
itself!