Revise shutdown coordination in gemini/aries-conduit
Commit changes the way the exitcode reduce is implemented when PSHM is
enabled to use shared memory within the node.  This is only marginally
faster (and more robust in face of shutdown-after-error) but is a
requirement for eliminating memory and GNI resource allocations used
for intra-node connections that are unused due to PSHM bypass on both
the AM and RDMA paths.