Eliminate allocation of GNI resources for intra-node/supernode comms.
The last few commits ensured that there are no more uses of GNI comms
for self communication (w/ or w/o PSHM), and no intra-supernode comms
when PSHM is in use.
So, this commit eliminates the corresponding GNI resources (EPs, CQ space
and calls to SmsgInit) and adds assertions to the Put, Get and AM paths.
A future commit will also reduce the Smsg mailbox memory allocation.