(Finally) replace the AM-based code for unaliagned gets GNI requires that Gets (both FMA and RDMA) be 4-byte aligned. This alignment restriction applies to the source and destination addresses and to the length of the Get. So, this code runs if any of src,dest,nbytes is not divisible by 4. The approach is to split the xfer into 3 pieces: head, tail and body.
Head is the first xfer, which uses bounce buffers (potentially the immediate
one) to "overfetch" such that we end on a 4-byte aligned source address.
However, if the source address is already properly aligned, Head is empty.

Tail is actually the second transfer, and is only non-empty if the length
remaining after Head is not a multiple of 4.  It xfers the "tail end" of
the address range (the last nbytes%128) using the immediate buffer.

After the head (which aligns the source) and the tail (which aligns the
length), the remaining "body" can be xfered in one of two ways, depending
on the alignment of the destination address:
1) If the dest address is now 4-byte aligned the we recursively call the
   normal get_nbi code which can use zero-copy if appropriate.
2) If the dest address is still unaligned, then we complete the xfer by
   chunking into one or more xfers that pass through bounce buffers and
   therefore get memcpy()ed into their unaligned final destination.

NOTE: I still need to explore a variation on the above that does the
"tail" only if the "body" is using the get_nbi path (#1), since the
alternative path (#2) is already trivially capable of handling an
unaligned length.