Implement "piggybacking" of credits.
The work committed last night "banks" credits that are not returned by
am explicit AM Reply, and sends an implicit reply only once for every
TWO such unreturned credits.  *This* commit goes the next step and
allows the next AM Request or Reply going to the proper peer to carry
the banked credit as a 1-bit field in the header (in addition to the
one credit implicitly carried by every AM Reply).

With just the coallescing, a Request/Request pattern (ping-pong or
flood) would send one implict Reply in each direction for every two
explicit Requests.  This is an average of 50% "overhead" (in msg count)
and these extra messages represent competition with the explicit
Request traffic headed in the same direction.

With the addition of piggybacking, the overhead of the Req/Req patterns
is reduced to ZERO, because the credits travel in the headers of the
explicit requests.  As a result, the 0-arg Short Request/Request
ping-pong case in testam now shows just a marginally higher (4%) latency
than Request/Reply, rather than 21% with only coallescing and 37% with
the naive code as is was yesterday morning.

Note that a one-way flood pattern still requires the "50% overhead" of
sending implicit Replies for every second Request.  However, in this case
the "extra" messages are travelling in the direction that is otherwise
silent (at least with respect to smsg traffic).  This could be lessed
by banking more than 1 credit, which is technically possible.  However,
the logic for getting that right (especially with pthreads) is too much
to deal with until I've had another cup of coffee (or two).  The pay-off
is likely small or even negative (due to introducing "bursty" behavior in
unidirectional flood patterns that would otherwise see the only benefit).
So, I am not planning to pursue this anytime soon.