Create a type of "control" message and use it for our non-AM messages.
This uses GNI_POST_CQWRITE to send up to 6 bytes of data to the same Cq
as the AM arrivals.  We need 23 bits for the source id, and 1 to distinguish
from the "normal" Cq entries.  That leaves us 3 bytes for a "control message"
which we currently split into an 8-byte OP and 16-bit ARG.  The hidden credit
return and shutdown messages are both now using this path.  The CqWrite is
cheaper than the mechanism we use of the full-blown AMs, in multiple ways.