Eliminate gasnetc_send_am() in favor of calling gasnetc_send_smsg().
This shifts the interface slightly to eliminate the need to ever copy an
AM header.  Previously, if the total length of the message to send was too
big for the 128-byte immediate buffer in the GPD, then the header would
need to be copied from the immediate buffer into a larger one before the
payload could be appended.  Now the total message length is examined BEFORE
the header is constructed, and the proper buffer is selected in gasnet_core.c
instead of in the (now removed) gasnetc_send_am() function.  Additionally
the appending of the payload has moved to gasnet_core.c where it is now
present ONLY in the Medium and packed-Long paths, as opposed to in all paths.
Note that agsnetc_send_am() was already an expensive no-op for Short AMs
and for non-packed Long AMs, because they are header-only and thus will
ALWAYS fit in the immediate buffer.