Continue even further reducing the overhead of AM reception
Following up on my previous commit, this one removes the per-peer
lock used when polling the AM mailboxes in a PAR build by arranging
that the "mb" operations take place entirely within the critical
section already present for protection of the AM recv queue.  That
only became possible w/ the previous commit that allowed the "release"
of the recv mailbox to always be performed before running the handler.