Commit an #if-0'ed change to the progress functions test.
This test has been observed to produce very long delays on some conduits
while finishing very quickly on others.  The cause has been determined to
be the fact that at each Poll the progress function under test will send
16 more AM Requests (which each require a Reply).  The result is that if
the "main" operations aren't completed quickly we can end up sending
millions (not an exageration) of AMs.

This change, still under consideration, will continue to send AMs in the
progress function, but will bound the number "in flight" (no Reply seen)
to some moderate value (128 in the current version).  While this does not
put any bound on the total number that will be sent, it does help at least
bound the number that are "buffered" somewhere in the system at any time.