First pass at reducing divergence between ibv/vapi conduit and ref-extended This commit is aimed at reducing *textual* differences in which the code ganerated should be identical (modulp quirks in compilers).
The eventual goal is to ease the comparision of the extended ref code and
vapi-conduit, in which eops have "binary" and "counted" completions,
respectively.  As part of the GASNet-EX work, I must decide to keep the
reference implementation as-is, change to counter-based, or implement
BOTH and allow the gasnet_extended_fwd.h select one or the other.

In both cases the iop is about 80+ bytes and includes the generality of
holding both Put and Get counters.
In the extended-ref code the eop is just 4 bytes, while the vapi-conduit
version of an eop is 12 bytes.
In the extended-ref code any gasnet-level op which requires multiple ops
at the lower level, one uses an iop via a recursive access region.
In vapi/ibv the need for multiple lower-level ops is relatively common,
and the identification of cases in which only a single op is required cannot
easily be done "up front".  For this reason, it is more space- and time-
efficient to use the counter-based eop consistently.

Note: the reference iop has padding between the non-atomic counters used
for initiated ops, and the atomic counters used to record the retired ops.
If one were to do the same for the vapi-conduit eop, the majority of the
space savings would probably be lost.  So, it is important that we get
some idea of how sensitve performance is to that padding.  My initial
expectation (yet to be tested):
+ For EOPs the initiated count grows only for a "short" period of time,
  unless there are lengthy stalls involved.  Only if another thread is
  polling during that initiation interval would there be a conflict on
  the cache line that contains the initiated and completed counters.
  And even then only if the first low-level OP is completed prior to
  the initiation of the last low-level OP.
  Thus the time interval in which the co-location of the two counters
  on the same cacheline leads to ping-ponging should be small.
+ For the IOPs the same counters are likely to be the subject of multiple
  initiation operations spread over time.  Therefore a given IOP could
  be subjected to more ping-ponging between caches than an EOP if one
  were to remove the padding.
+