Implement "unbounded" counters in vapi/ibv conduit.
This is a conduit-specific implementation of the work descibed (for the
extended reference implementation) in bug 1236 (Allow infinite number of
in-flight nbi operations).  As noted in that bug report, this is a valid
approach only because the underlying logic makes it impossible for the
counter "difference" (the number of in-flight internal ops) to get as
large as the range of an atomic_t.  In the vapi-conduit case the max
differece is bounded by GASNET_NETWORKDEPTH_TOTAL.

If there are no problems found with this code, the same logic will be
applied to the IOP code in the extended reference implementation and the
conduits which have cloned it.  That will be a little bit harder than
this work, because only vapi-conduit had abstracted the "increment".

With this commit the counters used by both eops and iops now run "forever"
without resetting to zero.  They just wrap at GASNETI_ATOMIC_MAX (with some
extra help for CPUs with atomics with less than the full range of a word).
This actually means LESS code on a critical path than before this commit.
The only exception is that in a TRACE/STATS build we reset the counters
in an IOP used for an NBI access region in order to have meaningfull values
reported when ending the region.