Work toward returning vapi/ibv-conduit to (nearly) the ref impl of IOP.
This commit separates the initiated and completed counters for the RDMA ops.
This will allow the IOP representation to be changed from the current use
of gasnetc_counter_t back to the use of initiated and completed counters on
distinct cache lines as in the reference implementaion.  Meanwhile, the eop
and internal counters for things like local completion remain unchanged.

Additional commits to restore the original iop/nbi logic will come later.