Next step toward unbounded NBI operations (bug 1236)
This commit factors the logic for testing completion of an IOP.
It includes the "& GASNETI_ATOMIC_MAX" that is required for "protection"
against wrap-around on CPUs with atomics implementations with a range
less than that of the full machine word.  For all "modern" CPUs that
will complile away to nothing when optimizations are enabled in any
decent compiler.

The wrap-protection is, at this point, still "unnecessary" because;
a) the code to reset the counters at 65000 is still present
b) comforming clients stil won't try for more than 65535 ops
However, removing (a) is the subject of the next commit.

Lapi-conduit is doing things just differently enough that I don't
want to mess with it.  So, it is not changed by this commit even
though it probably should be.