Significant simplifcation of gemini/aries-conduit extended API code.
By changing from the reference implementation of an eop to one using
initated and completed counters (inspired by, but not identical to,
vapi/ibv-conduit), the Get and Put code gets significantly simpler.
Previously, before we could launch any GNI-level ops for an explicit-
handle operation, we had to decide there would be more than one op
at the GNI-level.  If so, we had to allocate an iop in order to
get it's counter property.  Now, every eop has initiated_cnt and
completed_cnt fields.  Therefore we no longer need to "predict"
how many GNI-level operations are required before we can begin.
Instead we just increment the initiated_cnt for each GNI-level op.
The result is that there are significantly fewer branches in both
the initiation and completion paths.  The source code has also
shrunk significantly because much more code can be shared between
the NB and NBI paths since they are now much more similar.

I am looking to "merge" this back to extended-ref/gasnet_extended.c
such that a conduit author can have pre-written (#ifdef'd) code for
the counter-based implementation of an eop.  That is not 100%
certain, but will be worth while if it can be done without making
a total mess.