Restore OPSTATE checking to DEBUG builds of vapi/ibv and gemini/aries conduits.
This commit is "parallel" to my previous commit to extended-ref and the
other conduit-specific Extended API implementations, reducing the
differences between these two counter-based eop implementations and the
"boolean" version used in the other conduits.

Since these condits use a counter for defining completion of the eop,
there is no use of the COMPLETED state.  However, transitions into the
FREE and INFLIGHT states now match to those in the other conduits.

Not only does this reduce the diffs between the conduits, but it also
restores some useful consistency checking that had been removed in the
conversion from boolean-based to counter-based representation of the eop.