Split AM-based reference implementation
This commit removes from extended-ref/gasnet_extended.c all of the
gasnete_{get,put,memset}*() code and the AM handlers that support them.
These are moved to a distinct file extended-ref/gasnet_extended_amref.c
which is #included by extended-ref/gasnet_extended.c.
The names of the functions defined in gasnet_extended_amref.c are
prefixed with gasnete_amref_ instead of gasnete_, which allows them
to co-exist with conduit-specific implementations.
This will allow (soon) removal of the various cloned copies of the
AM-based Get/Put/Memset code that currently appear in several conduits.
It will also allow a near-trivial implementation of SEGMENT_EVERNYTHING
via use of the AM-based Extended API.