Commit "pmi-spawner" work
This commit factors the PMI support from, gemini and portals4 conduits.
The new code implements the compact ascii85 encoding for values used
in the PMP key-value-store (was a TODO item in portals4-conduit).
Additionally, this adds support for PMI-based launch to vapi, ibv and
mxm conduits (tested with SLURM's srun).