Add untested/unreachable code for strict conformance w/ latest docs.
When we have Edison back, we need to drop the "_xxxx" and actually test this.
The risk we face is that the inst id is obtained by a different path than
in the past.  In that case we may see the *true* remote instance ID of a
CQWRITE, rather than the fake one we've been encoding.  If so, then how we
distinguish CQWRITE events from normal (Put) events will require updating.