Remove conduit-specific per-thread handling in vapi/ibv conduit
The original development of vapi-conduit included adding a per-thread
struct in gasnet_core_sndrcv.c, which handles both AM and Extended API
operations.  This commit replaces that with use of the slot which the
extended-ref implementation already reserves for use by the core in its
per-thread data structure.

Tests on x86/vapi and x86-64/ibv show executable size got larger on x86
and smaller on x86-64 (though this could be due to different inlining).
The code performance was indistinguishable on both platforms, and of
course the source code is simpler.