Continued cleanups on vapi/ibv conduit Put / Get critical path
This commit follows-up on yesterday's work by eliminating the LOCAL address
from the gasnetc_rdma_{put,get}() functions in favor of putting it in the
sr_desc.  We also add assertions to ensure that both the remote and local
addresses in the sr_desc are advancing as expected (by comparision to the
src and dst values maintained only in a dbg build).