Further simplification of gasnetc_get_rkey_index()
This commit replaces lookup in a table (containing the ends of the
multiple registrations which comprise the segment) with a bit of
simple arithmetic.  The net change in this critical-path function
is equivalent to:

-  tmp = table[i] - offset;
+  tmp = var1 - (offset & var2);

where var1 and var2 are scalars, adjacent in memory to another
scalar used in the function (and thus likely cached together).
Var2 is new, but var1 existed before.

By making this replacement, we are able to eliminate the table
and the function that initializes it.