Another cleanup in vapi/ibv conduit
There is no need for
  #if GASNETC_FH_OPTIONAL
to control compilation of if()s containing !GASNETC_USE_FIREHOSE, because
when GASNETC_FH_OPTIONAL is 0, !GASNETC_USE_FIREHOSE is a pre-process time
constant 0 (and any decent optimizer will discard the unreachable code).
So, this is just a readabilty improvement, as the object code is
essentially unchanged.