Check for inline send only once per Put operation
This commit moves the check for inline sends outside of the "chunking"
loop in gasnetc_rdma_put().  By doing so both the small and large xfer
cases will involve fewer branches (and the intermediate will involve
no more than before).  The small xfer case completes without the loop
overhead, while the long case doesn't recheck "small enough for inline"
on every chunk.  This does eliminate the possible optimization of using
an inline send for the final chunk, but that is a relatively rare case.

This only became practical with the recent simplification (to the point
of elimination) of gasnetc_get_rkey_index().  Prior to that, the same
restructuring would have involved more duplicated code that would have
been unlikely to have been eliminated by the compiler's CSE pass.  Now,
however, the duplication is two simple expressions (for a total of just
four arithmetic operations).