Correction to a recent simplification
One of my earlier unintentionally relied on the fact that gcc's optimizer
is smart enough to discard the unreachable code.  However, that unreachable
code contains a call to a function which is not defined under exactly the
same circumstances as when the call is unreachable.  So, if the call is
NOT optimized away we'd experience a link failure.

This commit ensures we'll not see a link failure just an account of a
less than ideal optimizer.