Simplification of gasnetc_unpinned()
In this commit we take advantage of the fact that a client cannot construct
a valid call to a GASNet Put or Get in which the local buffer includes
portions both inside and outside the GASNet segment.

I intend to make that clear in the next rev of the documentation.