Simplify region bounds checks in gasnetc_zerocp_common()
This commit converts, as was done earlier with other code, the logic for
checking xfer ranges against the bounds of the seg_reg table (the table of
multiple registrations used to cover large GASNet segments).  The new logic
uses the shift and mask and therefore no longer needs the 'end' fields in
the seg_reg table.

With the removal of the last use of the 'end' field in the seg_reg table,
the field is being removed from the type gasnetc_memreg_t.  Its only
purpose in that struct was to simplify bounds checks like this one.