Don't initialize firehose in a singleton run
This is a dependency for another work-in-progress.
In that other work I discovered that for the singleton (gasneti_nodes==1)
case the firehose initialization was accessing an uninitilized 'snd_reg'
field in the HCA structure (which lead to a SEGV in that work-in-progress).
This commit skips the firehose_{init,fini}() calls for the singleton
case, since firehose is not used at all in such a run.