Reduce size of gasnet_memreg_t for ibv-conduit.
There is no need to store the lkey and rkey directly for ibv-conduit,
since they are part of the struct referenced by the "handle" field.
So, these two uint32_t fields are removed for ibv-conduit (but must
remain, for now, for vapi-conduit).

The added cost of the indirection is expected to be small, as the
"memory handle" is likely to be in-cache for peers we are actively
communicating with.  If that proves NOT to be the case, then we
should consider collecting just the lkeys into a dense array, as
is already done for rkeys.