Replace gasnetc_{hca,cep}_t.seg_reg[] with seg_lkeys[]
With the recent changes to how "seg_reg" bounds checking is done, we were
using just one member of the struct elements in hca->seg_reg[].  So, we can
save some space by discarding that array of structs in favor of an array of
just the one used member (the 'lkey').  This requires just 4 bytes per reg
instead of the previous 32.  We expect the number of these to be small
(1 to 4, or 10s at the most).  So, the new array is likely to fit in a
single cache line (and we allocate cache-aligned memory for it).

Since another recent change re-duplicated the lkey in gasnetc_memreg_t (also
to save space), this commit also has the added benefit of removing the level
of indirection that was needed to reach the lkey via the seg_reg[] array.
It was actually while looking into *that* issue that I discovered that the
'lkey' was the only member still used in hca->seg_reg[].