Modify gasnetc_pin_maxsz processing to favor pinning one large region.
The original VAPI firmware had a serious performance defect when one
had single memory registrations larger than about 256K.  So, the code
for vapi-conduit was originally written to favor pinning the GASNet
segment as multiple smaller registrations (length <= gasnetc_pin_maxsz).
At the time we decided to be "clever" by additionally ensuring that
gasnetc_pin_maxsz was never larger than the HCAs max message size (1G
is typical).  Thus we only needed a single bounds check in the main
"chunking" loop.

When IBV support was added we made no changes to the logic described
above, but replaced the default value with the max message size instead
of the 256K value that was meant to address a VAPI defect (not present
in the IBV firmware).

This commit modifies how gasnetc_pin_maxsz and some related values are
setup for ibv-conduit (vapi-conduit's default behavior is unchanged).
Now if it is possible to register the segment with a single region, then
that is done and the gasnetc_pin_maxsz* variables are setup such that the
bounds checking logic "just works" to enforce the max message size.  The
"trick" here is that there are no extra branches required.  In fact, the
only changes in gasnet_core_sndrcv.c are updates to comments - the main
"chunking loop" did not require any changes for this to work.

As a result, we are now able to initiate xfers up to this maximum size
WITHOUT any need to artificially align to region boundaries.  For an HCA
with a max message size of 1G, a collection of uniformly randomly aligned
1G GASNet-level xfers would have been performed using 2 ibv-level xfers
for all but the perfectly aligned case.  In fact, because BOTH the local
and remote segments were registered in multiple regions, a 1G xfer with
local data in-segment would typically require THREE distinct xfers.  Now,
1G gasnet-level xfers can always be performed as single ibv-level xfers.