Revise the throttled AMPoll implementation
The previous implementation was written prior to the introduction of
progress functions.  When they were added, they were just tacked on
to the end.  However, this means that if other pthreads are constantly
doing "useful work", then progress functions could be starved.

This commit ensures that progress functions cannot be starved by the
active senders.