Fix some compilation errors for pami-conduit with PSHM enabled.
Clearly this code had not been tested before, and as of this commit it
STILL has not been tested because PSHM is not yet functional on BG/Q,
which is my only access to PAMI at the moment.

I am actively pursuing getting PSHM working on BG/Q in order the ensure
that this code gets tested.  However, the worst case for the current
commit is that I've moved the failure from compile-time to run-time.