Fixes and additions to get PSHM-over-POSIX working on BG/Q
In addition to the idiosyncrasies seen on BG/P, there is one more
to deal with on BG/Q.  If one close()s the fd obtained from the
shm_open() call (as one normally would) then the kernel appears to
eventually reuse the memory.  Though I've not tracked down the
details, I found that the memory of the first shm_open() was getting
zeroed by a latter mmap() call (though the address ranges did NOT
overlap).