Improved PSHM support on BG/Q
The commit of last night was sufficient to get correct operation, but
was leaking file desciptors and additionally was giving too-high estimates
for the max segment size.

This commit implements (naively) a table for mapping mmap() address to
the corresponding file descriptor, and does the close() call at unmap
time.  Not only does this resolve the descriptor leak, but it also
(mysteriously) fixed the over-estimation of max segment size.