Fix an issue with the multi-threaded case.  If one thread is processing the last message in a buffer while another thread starts processing the AUTO_UNLINK, there's a race between reposting the buffer and reading the data out of the last message.
Our solution is to have a counting event attached to the ME associated
with the buffer and an atomic for each buffer which is incremented at
the end of a message processing loop.  When the AUTO_UNLINK shows up,
the thread handling the UNLINK waits until the atomic and the counting
event are the same.
CVSMAIL: CC: