Merge Khaled's multi-domain work to HEAD (disabled by default)
This commit merges the branch MDOM_GNI_09122013 to HEAD.

At the point just prior the the merge:
   + MDOM_GNI_09122013 is tagged MDOM_GNI_09122013_to_HEAD
   + HEAD is tagged pre_merge_MDOM_GNI_09122013
Immediately after this commit I will tag HEAD with post_merge_MDOM_GNI_09122013
So, MDOM_GNI_09122013_to_HEAD and post_merge_MDOM_GNI_09122013 will
be identical except for RCS ident strings.
Tags are on files (potentially) modified on the branch, NOT the entire repo.

If necessary the following can be used to revert this merge:
  cvs up -kk -j post_merge_MDOM_GNI_09122013 -kk -j pre_merge_MDOM_GNI_09122013 -kk
(though one or more of those "-kk" is probably unnecessary).