Update docs and run script for PSHM on BG/Q
It turns out setting BG_MAPCOMMONHEAP=1 (and exporting to the jobs) is
critical to proper PSHM support.  Without this setting the entire shared
memory (up to 64 shared heaps) are all carved out of the address space
of the FIRST process on each node.  This applies to both PAMI and MPI.