Merge BUG3178_branch to HEAD
This commit makes HEAD the same as the tip of BUG3178_branch.
In addition to including the fix for bug 3178, this commit also
includes several multi-domain related commits which were made
to that branch to reduce the testing required in the limited time
remaining before release.  These changes include a merge of the
logic Khaled committed on the branch GEMINI_LONGAM_PAYLOAD_16092013.

As for the fix for bug 3178:

The core problem is that GNI will not permit memory registration of
a "foreign" segment.  This is sensible because those pages were imported
into our address space using the XPMEM mechanism.  Unlike SysV or
POSIX shared memory, XPMEM does not outlive the creator.  Therefore,
if GNI were to allow memory registration of physical page frames that
were later removed from our virtual address space (WITHOUT any way to
notify the GNI driver or firmware) then the NIC might continue to
read or write to those page frames after then have been allocated to
another process (or within the kernel).  Unfortunately, Cray has
told us XPMEM is *the* shared memory mechanism.  I hope to look
deeper when I can to see if POSIX shared memory might be present
too (it certainly wasn't when we began GNI support).  However, the
release is too near to make that sort of a change.

The approach taken by the bug fix had four parts:

1) The dynamic memory registration path in gasnet_gemini.c now has
an int (rather than void) return type, which indicates success or
failure of the registration request.  Previously we would just
gasneti_fatalerror() if the registration failed.

2) The Put and Get code paths requesting dynamic memory registration
check the new return value and IF it indicates failure then they
both fall back on performing a PARTIAL xfer usign the maximum-sized
bounce-buffer (note that there would not have been any attempt to
dynamicaly register any memory had the orignal request fit in one
bounce-buffer).  These two paths now return the size of the xfer
(instead of void) which is "full" except in this exceptional case
of the failed dynamic memory registration.

3) The gasnet_extened.c code for Put and Get already had a "chunking"
loop that would split the request into chunks that were valid for
the in-vs-out of segment choice.  In particular, up to a little under
2G for in-segment and 1M for out-of-segment (where 1M is the empirically
determined "safe" limit for dynamic memory registration).  The loops
have been modified slightly to take the return value of the lower-level
xfer call as the size for the chunks.  So, in the absence of the bug
the loops use the orignally chosen chunk size and nothing has really
changed.  However, in the case of a fall-back to bounce-buffers not
only does the returned value ensure that src, dst and nbytes advance
by the proper values, but it also ensures that any subsequent chunks
will be equal to the maximm bounce-buffer size and thus will NOT have
the overhead of any additional failed dynamic memory registrations (for
the extent of the current GASNet-level Put or Get).  [Note: the loops
that perform chunking follow a preamble that tries to ensure that
xfers larger than 2 chunks will page align the second and subsequent
ones.  That had to change slightly more than the description above
might suggest, but the new code is not radically different]

4) The code for moving Long payloads too large to pack like a Medium
now includes a loop to react to the "short write" return values as
needed to ensure that all of the payload is sent.  This is relatively
straight-forward for the Long, which blocks for the payload Put(s)
before it send the headers.  However, the LongAsync does a trick to
"chain" the header send after payload Put.  That is stil the case,
except that now the payload Put may consist of multiple descriptors,
all of which will reference the header descriptor as "next".  The
addition of a counter in the otherwise unused "immediate" space in
the header descriptor ensures that one can unambiguously identify
the LAST portion of the Put to complete and only then issue the
header send.


No caching or other mechanism as been implemented to "learn" beyond
the bounds of one GASNet-level Put or Get operation.  This is done
under the belief that this error-recovery path is a relatively un-
common case and that we don't want to burden the normal case.  That
belief is born-out by the fact that we'd never seenteh problem before
this past couple of week.